requirements.txt
torch
torchaudio
librosa
matplotlib
numpy>=1.22
omegaconf
pandas
hydra-core
jiwer
tqdm
openai-whisper
librosa
soundfile